requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-multipart
aiofiles
pillow
